調整

+++ tests/test_subtitle.py
@@ -1,10 +1,11 @@
 import requests
 from moviepy import *
 import re
+import os
 
 
 # 字幕クリップの作成
-def make_subtitle_clip(text):
+def make_subtitle_clip(text,current_time):
     subtitle_font = "fonts/Corporate-Logo-Rounded-Bold-ver3.otf"
     subtitle_font_size = 24
     subtitle_color = 'white'
@@ -12,15 +13,15 @@ def make_subtitle_clip(text):
     subtitle_stroke_width = 2
     speaker = 1
     speed = 1
-    silence_duration = 1  # 音声合成の間の無音時間（秒）
+    silence_duration = 0.5  # 音声合成の間の無音時間（秒）
 
     # 音声合成の関数を呼び出して音声クリップを作成
-    voice_clip = make_voice_clip(text, speaker=speaker, speed=speed)
+    voice_clip = make_voice_clip(text, speaker=speaker, speed=speed, silence_duration=silence_duration)
     clip_duration = voice_clip.duration
     
     # テキストクリップを作成
     subtitle_clip = TextClip(text=text, font=subtitle_font, font_size=subtitle_font_size, color=subtitle_color, stroke_color=subtitle_stroke_color, stroke_width=subtitle_stroke_width, size=(1700, 100), method='caption')
-    subtitle_clip = subtitle_clip.with_position(('center', 'bottom')).with_duration(clip_duration+silence_duration)
+    subtitle_clip = subtitle_clip.with_position(('center', 'bottom')).with_duration(clip_duration).with_start(0)
     video_clip = CompositeVideoClip([subtitle_clip])
     video_clip = video_clip.with_audio(voice_clip)
     
@@ -31,31 +32,31 @@ def make_voice_clip(text, speaker=1, speed=1, silence_duration=1):
     filename = re.sub(r'[\\/*?:"<>|]', "_", text)
     filepath = f"data/temp_data/voice_{filename}.wav"
     
-    
-    # 1. テキストから音声合成のためのクエリを作成
-    query_payload = {'text': text, 'speaker': speaker}
-    query_response = requests.post(f'http://localhost:50021/audio_query', params=query_payload)
+    if not os.path.exists(filepath):
+        # 1. テキストから音声合成のためのクエリを作成
+        query_payload = {'text': text, 'speaker': speaker}
+        query_response = requests.post(f'http://localhost:50021/audio_query', params=query_payload)
 
-    if query_response.status_code != 200:
-        print(f"Error in audio_query: {query_response.text}")
-        return
+        if query_response.status_code != 200:
+            print(f"Error in audio_query: {query_response.text}")
+            return
 
-    query = query_response.json()
-    #トークスピードを設定
-    query["speedScale"] = speed
+        query = query_response.json()
+        #トークスピードを設定
+        query["speedScale"] = speed
 
-    # 2. クエリを元に音声データを生成
-    synthesis_payload = {'speaker': speaker}
-    synthesis_response = requests.post(f'http://localhost:50021/synthesis', params=synthesis_payload, json=query)
+        # 2. クエリを元に音声データを生成
+        synthesis_payload = {'speaker': speaker}
+        synthesis_response = requests.post(f'http://localhost:50021/synthesis', params=synthesis_payload, json=query)
 
-    if synthesis_response.status_code == 200:
-        # 音声ファイルとして保存
-        with open(filepath, 'wb') as f:
-            f.write(synthesis_response.content)
-        print(f"音声が {filepath} に保存されました。")
-    else:
-        print(f"Error in synthesis: {synthesis_response.text}")
-        return
+        if synthesis_response.status_code == 200:
+            # 音声ファイルとして保存
+            with open(filepath, 'wb') as f:
+                f.write(synthesis_response.content)
+            print(f"音声が {filepath} に保存されました。")
+        else:
+            print(f"Error in synthesis: {synthesis_response.text}")
+            return
     
     # 音声クリップを作成
     audio_clips = []
